include the title of datasets
This is to prevent clients from having to fetch the dataset URLs to get overview/list data. Could be expanded to the spatial/temporal extent. 
Note that I left out the @type here (dcat:Dataset). I think it is not necessary here as this is dictated by the ontology anyway as range type, but also because it appears in the actual detail dataset URL. This is a question of how clients read the (RDF) information, how clever they are, how they harvest etc. I am not sure about that currently.